requirements.txt
# Universal News - Python Dependencies
# Step 1: Core Agent Framework

# Core Framework
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM Integration
openai>=1.6.0

# Async Processing
asyncio>=3.4.3

# NLP (for future steps)
# spacy>=3.7.0
# sentence-transformers>=2.2.0
# scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
